refactor: move profile items to composable
